+++ pcs/cli/routing/pcsd.py
@@ -1,4 +1,5 @@
 from pcs import pcsd, usage
+from pcs.cli.common.errors import raise_command_removed
 from pcs.cli.common.routing import create_router
 
 pcsd_cmd = create_router(
@@ -6,9 +7,13 @@ pcsd_cmd = create_router(
         "help": lambda lib, argv, modifiers: print(usage.pcsd(argv)),
         "accept_token": pcsd.accept_token_cmd,
         "deauth": pcsd.pcsd_deauth,
-        "certkey": pcsd.pcsd_certkey_cmd,
+        "certkey": lambda lib, argv, modifiers: raise_command_removed(
+            pcs_version="1.0"
+        ),
         "status": pcsd.pcsd_status_cmd,
-        "sync-certificates": pcsd.pcsd_sync_certs,
+        "sync-certificates": lambda lib, argv, modifiers: raise_command_removed(
+            pcs_version="1.0"
+        ),
     },
     ["pcsd"],
 )