+++ pcsd/pcsd.rb
@@ -195,10 +195,6 @@ post '/run_pcs' do
       'only_superuser' => false,
       'permissions' => nil,
     },
-    ['pcsd', 'sync-certificates', '...'] => {
-      'only_superuser' => false,
-      'permissions' => Permissions::FULL,
-    },
     ['quorum', 'device', 'status', '...'] => {
       'only_superuser' => false,
       'permissions' => Permissions::READ,